feat(server): extract serveFromBackend to BaseService
